fix path

+++ init.sh
@@ -1,6 +1,21 @@
-ln -s $HOME/dotfiles/vimrc $HOME/.vimrc
-ln -s $HOME/dotfiles/zshrc $HOME/.zshrc
-ln -s $HOME/dotfiles/zsh $HOME/.zsh
+if [ ! -e $HOME/.vimrc ]; then
+    ln -s $HOME/dotfiles/vimrc $HOME/.vimrc
+fi
 
-curl https://raw.githubusercontent.com/Shougo/dein.vim/master/bin/installer.sh > installer.sh
-sh ./installer.sh ~/.vim
+if [ ! -e $HOME/.zshrc ]; then
+    ln -s $HOME/dotfiles/zshrc $HOME/.zshrc
+fi
+
+if [ ! -e $HOME/.zsh ]; then
+    ln -s $HOME/dotfiles/zsh $HOME/.zsh
+fi
+
+if [ ! -e $HOME/.vimperatorrc ]; then
+    ln -s $HOME/dotfiles/vimperatorrc $HOME/.vimperatorrc
+fi
+
+if [ ! -e $HOME/.vim/repos/github.com/Shougo/dein.vim ]; then
+    curl https://raw.githubusercontent.com/Shougo/dein.vim/master/bin/installer.sh > installer.sh
+    sh ./installer.sh ~/.vim
+fi
+
